+++ src/kani/monde/Lieu.java
@@ -32,12 +32,12 @@ void dessiner(float X, float Y)  {
 
 
     @Override
-    int getX() {
+   float getX() {
     return posx;
 
     }
 
-    int getY() {
+  float getY() {
     return posy;
     }
